Added files via upload
A hexagon created from MS PowerPoint for use as a custom shape in tableau. Create a new folder under the "My Tableau Respositry/Shapes" directory on your computer as "Maps" and save this png file there. Then from Tableau, click on Shapes and reload shapes. The "Maps" folder will appear as an option to assign an icon and the hex icon will be available. T